+++ std/compress/lzss/decompress_test.go
@@ -126,36 +126,6 @@ func testCompressionRoundTripSnark(t *testing.T, d, dict []byte, options ...test
 	test.NewAssert(t).CheckCircuit(circuit, test.WithValidAssignment(assignment), test.WithBackends(backend.PLONK), test.WithCurves(ecc.BN254))
 }
 
-func TestReadBytes(t *testing.T) {
-	expected := []byte{254, 0, 0, 0}
-	circuit := &readBytesCircuit{
-		Words:      make([]frontend.Variable, 8*len(expected)),
-		WordNbBits: 1,
-		Expected:   expected,
-	}
-	words, err := goCompress.NewStream(expected, 8)
-	assert.NoError(t, err)
-	words = words.BreakUp(2)
-	assignment := &readBytesCircuit{
-		Words: test_vector_utils.ToVariableSlice(words.D),
-	}
-	test.NewAssert(t).CheckCircuit(circuit, test.WithValidAssignment(assignment), test.WithBackends(backend.PLONK), test.WithCurves(ecc.BN254))
-}
-
-type readBytesCircuit struct {
-	Words      []frontend.Variable
-	WordNbBits int
-	Expected   []byte
-}
-
-func (c *readBytesCircuit) Define(api frontend.API) error {
-	byts := combineIntoBytes(api, c.Words, c.WordNbBits)
-	for i := range c.Expected {
-		api.AssertIsEqual(c.Expected[i], byts[i*8])
-	}
-	return nil
-}
-
 func getDictionary() []byte {
 	d, err := os.ReadFile("./testdata/dict_naive")
 	if err != nil {
