+++ std/compress/lzss/testing.go
